+++ lib/ex_athena/tools/read.ex
@@ -39,6 +39,9 @@ defmodule ExAthena.Tools.Read do
   end
 
   @impl true
+  def parallel_safe?, do: true
+
+  @impl true
   def execute(args, %ToolContext{} = ctx) do
     with {:ok, path} <- fetch_path(args, ctx),
          {:ok, stat} <- File.stat(path),